backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-ai
python-dotenv
# google-genai (Using pydantic-ai's GeminiModel instead)